Added Salary-Slip generation functionality

+++ server/db/migrations/20240426113151-create-user.js
@@ -90,7 +90,7 @@ module.exports = {
         allowNull: false,
       },
       availableLeaves: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       commissionFlag: {
@@ -98,7 +98,7 @@ module.exports = {
         allowNull: false,
       },
       commissionPercentage: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL,
         allowNull: false,
       },
       createdAt: {
